Extension cleanup and disposable handling
